Uzbek translation
This is my first time contributing to an Uzbek translation, and I'm hopeful that others will join in to refine and enhance this work. As of now, I'm the only maintainer working on this, but I welcome others who are interested in collaborating to ensure a well-rounded translation. If anyone is interested, please feel free to comment below so we can maintain this together as a team.