Convert to nextjs app
This is a simple one page application that fetches data from GH in a useEffect.

We can leverage a server function to handle the search and caching and then SSR the page content. This change just converts the app as is to use Next. The next commits restructure the app to take advantage of server rendering.