Convert to NextJS app (#350)
* Convert to nextjs app

This is a simple one page application that fetches data from GH in a useEffect.

We can leverage a server function to handle the search and caching and then SSR the page content. This change just converts the app as is to use Next. The next commits restructure the app to take advantage of server rendering.

* Convert glamor to css modules

Glamor is client-only which means we'd have to add the 'use client' directive everywhere that we style. Let's instead switch to simple CSS modules to enable SSR.

* Move GH search request to server

Instead of rendering empty and then searching in an effect, we can move the search request to the server to SSR the complete site.

* Update dependencies and fix eslint

Update to latest while we're here and make sure linting works with next integration

* Update UI styling

Add some polish to the UI while we're here updating everything else.

* Add error handling for failed GH search

* Remove sort selector weight change

* Fix deployment